+++ src/i18n/langs/cn.json
@@ -739,6 +739,7 @@
     "communityEdition": "社区版",
     "consult": "咨询",
     "createUserSetting": "创建用户设置",
+    "wechatSetting": "企业微信认证设置",
     "emailCustomUserCreatedBody": "创建用户邮件的内容",
     "emailCustomUserCreatedHonorific": "创建用户邮件的敬语",
     "emailCustomUserCreatedSignature": "署名",
