+++ src/i18n/langs/en.json
@@ -732,6 +732,7 @@
     "communityEdition": "Community edition",
     "consult": "Consult",
     "createUserSetting": "Create User setting",
+    "wechatSetting": "Wechat setting",
     "emailCustomUserCreatedBody": "Create user email content",
     "emailCustomUserCreatedHonorific": "Create user honorific",
     "emailCustomUserCreatedSignature": "Signature",
